topgun-back/requirements.txt
django~=4.2.1
djangorestframework~=3.14.0
djangorestframework-jwt~=1.11.0